refactor: moving actor tests
